Construction of 2D grid
Now the 2D percolation lattice is up and running. It can be modified to have the water source from all sides or the center, and now after a sufficient number of frames the animation will stop and not crash my computer.

I have made the file for converting the 2d matrix into the first layer of the neural network, which seems straighforward given background research.

Current issues with existing code:
-Chalenges saving animations after percolation simulation
-when the percolation sim is called as a function (done to streamline things) the animation/simulation is not working yet, despite the fact that the simualtion works when the lines of the function are called globally.