Improved Percolation and Started network
In my notebooks I have made progress.

I have added the returns to the perclation sim function so that it will return the percentage submerged for the groundwater simulation. It also returns a numpy array that contains the initial conditons of the lattice.

For the neural network setup I have been doing a lot of reading/research and treating the python notebook as a lab notebook/starting point for my presentation of the math behind training the network.

In terms of code I have functions for the way that the nodes will process the inputs to make their outputs, as well as the Network class defined. Beyond that I need to finish getting through the math to finish coding the network, but I anticipate doing that this weekend.